nit(archival): extract Zod schemas, add runtime parse + error log to fetchArchivedExecution
