+++ packages/backend/src/helpers/archival/fetch-archived-execution.ts
@@ -9,8 +9,7 @@ import { gunzip } from 'node:zlib'
 import logger from '../logger'
 
 import { S3_PREFIX_EXECUTIONS, S3_PREFIX_TEST_EXECUTIONS } from './build-s3-key'
-import { ArchivedPayloadSchema } from './schemas'
-import type { ArchivedPayload } from './types'
+import { type ArchivedPayload, ArchivedPayloadSchema } from './types'
 
 const gunzipAsync = promisify(gunzip)
